% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrchart.r
\name{imrchart}
\alias{imrchart}
\title{A function to create individuals and moving range charts}
\usage{
imrchart(data, xcol = NULL, ycol = NULL, bMed = F, bSD = F,
  bTests = rep(T, 8), pTs = c(2, 4, 6, 9, 8, 14, 15))
}
\arguments{
\item{data}{data frame or time series containing the data to be plotted}

\item{xcol}{sequence data, default is a sequence called "Index"}

\item{ycol}{process output data}

\item{bMed}{a logical statement, T = control limits based on median moving range, F = control limits based on mean moving range, defaults to FALSE}

\item{bSD}{a logical parameter, T = show standard deviation lines, defaults to F}

\item{bTests}{a logical parameter vector of length 8, T = show test number n, defautls to rep(T,8)}

\item{pTs}{a vector of integers for the test parameter, defaults to c(2,4,6,9,8,14,15)}
}
\description{
This function allows you to create an individuals and moving range chart from a single, numeric data vector
}
\examples{
imrchart(discoveries,bTests = c(T,T,rep(F,6)))
}
\keyword{chart,}
\keyword{control}
\keyword{imr,}
\keyword{ymr}

