% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imrchart.r
\name{imrchart}
\alias{imrchart}
\title{A function to create individuals and moving range charts}
\usage{
imrchart(x, xname, xLabName, bMed = F, bSD = F, bTests = rep(T, 8),
  pTs = c(2, 4, 6, 9, 8, 14, 15))
}
\arguments{
\item{x}{a numeric data vector, no default}

\item{xname}{a character string for the name of the data vector, no default}

\item{xLabName}{a character string for the name of the x-axis, defaults to 'Index'}

\item{bMed}{a logical statement, T = control limits based on median moving range, F = control limits based on mean moving range, defaults to FALSE}

\item{bSD}{a logical parameter, T = show standard deviation lines, defaults to F}

\item{bTests}{a logical parameter vector of length 8, T = show test number n, defautls to rep(T,8)}

\item{pTs}{a vector of integers for the test parameter, defaults to c(2,4,6,9,8,14,15)}
}
\description{
This function allows you to create an individuals and moving range chart from a single, numeric data vector
}
\examples{
imrchart()
}
\keyword{chart,}
\keyword{control}
\keyword{imr,}
\keyword{xmr}
